Assignment Question 04
Users can put a compressed file at any path of the Linux file system. The name of the file will be research and the extension will be of compression type, example for gzip type extension will be .gz. You have to find the file and check the compression type and uncompress it.